refactor: ObjectsMap split on parts
